test war win

+++ tests/test.cpp
@@ -18,11 +18,11 @@ Json object("{\n"
             "}");
 EXPECT_EQ(std::any_cast<std::string>(object["lastname"]), "Ivanov");
 EXPECT_EQ(std::any_cast<bool>(object["islegal"]), false);
-EXPECT_EQ(std::any_cast<unsigned>(object["age"]), 25);
+EXPECT_EQ(std::any_cast<double>(object["age"]), 25);
 
 auto marks = std::any_cast<Json>(object["marks"]);
-EXPECT_EQ(std::any_cast<unsigned>(marks[0]), 4);
-EXPECT_EQ(std::any_cast<unsigned>(marks[1]), 5);
+EXPECT_EQ(std::any_cast<double>(marks[0]), 4);
+EXPECT_EQ(std::any_cast<double>(marks[1]), 5);
 
 auto address = std::any_cast<Json>(object["address"]);
 EXPECT_EQ(std::any_cast<std::string>(address["city"]), "Moscow");
